Move and remove e2e methods
