+++ Gopkg.lock
@@ -1427,7 +1427,6 @@
     "github.com/docker/cli/cli/context/store",
     "github.com/docker/cli/cli/flags",
     "github.com/docker/cli/opts",
-    "github.com/docker/cnab-to-oci/converter",
     "github.com/docker/cnab-to-oci/remotes",
     "github.com/docker/distribution/reference",
     "github.com/docker/docker/api/types",
